(2022-07-26) chat-service repository test 단위 테스트로 리팩토링

+++ chat-service/src/main/java/webrtc/chatservice/controller/HttpApiController.java
@@ -58,7 +58,7 @@ public class HttpApiController {
 
     public void postDecreaseUserPoint(String email, Long point) {
         try {
-            ResponseEntity<String> response = postRequest("http://auth-service:8080/api/v1/webrtc/auth/decrease/point", new DecreasePointRequest(email, point));
+            ResponseEntity<String> response = postRequest("http://auth-service:8080/api/v1/webrtc/auth/decrease/point", new DecreasePointRequest(email, point.intValue()));
         } catch (HttpClientErrorException e) {
             throw new InsufficientPointException();
         }
